add a task parser and an indentation counter for the task aggregator
Alright, I have the basic components down that I need as well as some
quick tests for them to ensure they at least work in the cases I need them
too. Next step is building a task file to object hierarchy converter and
then an aggregator to build out some stats. Once that is done I'll just
need a way to output the results. I kinda like the idea of them being
tasks with a count of the times they appeared in their state (checked vs
unchecked). Might be clearer if I just use the text of the checkboxes, and
then then just added a small summary in paranthesis like (# checked, #
unchecked) if the count is above 1